+++ parsley-core/src/ghc-8.10+/Parsley/Internal/Backend/Machine/InputOps.hs
@@ -84,10 +84,9 @@ instance InputPrep CharList where
   prepare qinput = [||
       let CharList input = $$qinput
           next (# i#, c:cs #) = (# c, (# i# +# 1#, cs #) #)
-          I# size# = length input
-          more (# i#, _ #) = $$(intLess [||i#||] [||size#||])
-          --more (# _, [] #) = False
-          --more _            = True
+          more :: (# Int#, [Char] #) -> Bool
+          more (# _, [] #) = False
+          more _           = True
       in (# next, more, $$(offWith [||input||]) #)
     ||]