+++ parsley/benchmarks/BrainfuckBench/Main.hs
@@ -50,16 +50,16 @@ brainfuck =
   let bfTest :: NFData rep => (FilePath -> IO rep) -> String -> (rep -> Maybe [BrainFuckOp]) -> Benchmark
       bfTest = benchmark ["benchmarks/inputs/helloworld.bf", "benchmarks/inputs/helloworld_golfed.bf", "benchmarks/inputs/compiler.bf"]
   in bgroup "Brainfuck"
-       [ {-bfTest string          "Parsley (Stream)"          (brainfuckParsleySS . CharList)
-       ,-} bfTest string          "Happy"                     BrainfuckBench.Happy.Parser.brainfuck
+       [ bfTest string          "Parsley (Stream)"          (brainfuckParsleySS . CharList)
        , bfTest string          "Parsley (String)"          brainfuckParsleyS
        , bfTest text            "Parsley (Text)"            (brainfuckParsleyT . Text16)
-       --, bfTest bytestring      "Parsley (ByteString)"      brainfuckParsleyB
+       , bfTest bytestring      "Parsley (ByteString)"      brainfuckParsleyB
        --, bfTest lazy_bytestring "Parsley (Lazy ByteString)" brainfuckParsleyLB
+       , bfTest string          "Handrolled"                BrainfuckBench.Handrolled.Parser.brainfuck
+       , bfTest string          "Happy"                     BrainfuckBench.Happy.Parser.brainfuck
        , bfTest string          "Parsec (String)"           (parsecParse BrainfuckBench.Parsec.Parser.brainfuck)
        , bfTest text            "Parsec (Text)"             (parsecParse BrainfuckBench.Parsec.Parser.brainfuck)
        , bfTest string          "Mega (String)"             (megaParse BrainfuckBench.Megaparsec.Parser.brainfuck)
        , bfTest text            "Mega (Text)"               (megaParse BrainfuckBench.Megaparsec.Parser.brainfuck)
        , bfTest text            "Atto (Text)"               (attoParse BrainfuckBench.Attoparsec.Parser.brainfuck)
-       , bfTest string          "Handrolled"               BrainfuckBench.Handrolled.Parser.brainfuck
        ]